+++ examples/runtime.rs
@@ -5,7 +5,7 @@ use log::{debug, error, info};
 use pretty_env_logger;
 use tokio::{self, runtime::Runtime};
 
-use reool::{Commands, RedisPool};
+use reool::*;
 
 /// Do many ping commands with no checkout timeout
 /// and an unbounded checkout queue runing on the default runtime
@@ -20,7 +20,7 @@ fn main() {
             .connect_to_node("redis://127.0.0.1:6379")
             .desired_pool_size(10)
             .reservation_limit(None) // No limit
-            .checkout_timeout(None) // No timeout
+            .checkout_mode(None) // No timeout
             //.task_executor(runtime.executor()) no explicit executor!
             .finish_redis_rs()
             .unwrap();
@@ -28,7 +28,7 @@ fn main() {
         info!("Do one 1000 pings concurrently");
         let futs: Vec<_> = (0..1_000)
             .map(|i| {
-                pool.check_out()
+                pool.check_out(PoolDefault)
                     .from_err()
                     .and_then(Commands::ping)
                     .then(move |res| match res {